bec/api/views.py
from rest_framework import generics
from .models import Project, Lead
from .serializers import ProjectSerializer, LeadSerializer

class ProjectListView(generics.ListAPIView):
    queryset = Project.objects.all()
    serializer_class = ProjectSerializer

class LeadCreateView(generics.CreateAPIView):
    queryset = Lead.objects.all()
    serializer_class = LeadSerializer